requirements.txt
transformers
streamlit
pillow
torch
requests
